refactor: reorg BehaviourDB.lua to HotAC/AI
